requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
scikit-learn
google-generativeai
python-dotenv
numpy
requests
python-docx
